+++ packages/playground-website/src/config.ts
@@ -1,0 +1,16 @@
+import samples from "../samples/dist/samples.js";
+
+export const TypeSpecPlaygroundConfig = {
+  defaultEmitter: "@typespec/openapi3",
+  libraries: [
+    "@typespec/compiler",
+    "@typespec/http",
+    "@typespec/rest",
+    "@typespec/openapi",
+    "@typespec/versioning",
+    "@typespec/openapi3",
+    "@typespec/json-schema",
+    "@typespec/protobuf",
+  ],
+  samples,
+} as const;
