+++ packages/playground-website/src/index.ts
@@ -1,16 +1,2 @@
-import samples from "../samples/dist/samples.js";
-
-export const TypeSpecPlaygroundConfig = {
-  defaultEmitter: "@typespec/openapi3",
-  libraries: [
-    "@typespec/compiler",
-    "@typespec/http",
-    "@typespec/rest",
-    "@typespec/openapi",
-    "@typespec/versioning",
-    "@typespec/openapi3",
-    "@typespec/json-schema",
-    "@typespec/protobuf",
-  ],
-  samples,
-} as const;
+export { TypeSpecPlaygroundConfig } from "./config.js";
+export { ImportToolbarButton } from "./import-openapi3.js";
